+++ elmyra.ip.access.epo/elmyra/web/identity/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+# (c) 2014 Andreas Motl, Elmyra UG
+
+# ------------------------------------------
+#   bootstrapping
+# ------------------------------------------
+def includeme(config):
+    config.include(".store")
+    config.scan(".service")